move time functions to baseclass
